wrap resize component listener to react hook
