+++ static/src/hooks/use-resize-component/index.js
@@ -1,0 +1,1 @@
+export { default as useResizeComponent } from './use-resize-component'
